refactor(session): extract useOptimizer from DesignSession
The whole reactive-optimiser cluster moves into
components/session/useOptimizer.ts: the state, the unconditional render-time
optEnabledRef mirror (still a render write, not an effect), the design-load
reset effect, runOptimize, the optFixedSig memo and its reactive effect, the
pause-cue effect, knobOptFor/updateKnobOpt and the knob-menu Escape effect.
Bodies and all five literal dep arrays are unchanged.

optAbortRef and optEnabledRef are returned raw because selectVariant and
handleUserParamChange stay behind and mutate them directly. buildRequest and
setParamAtPath are passed as plain per-render values, not memoized.

The hook is called at optFixedSig's old position, so the cluster's three
trailing effects keep their global order exactly; only the design-load reset
moves, from first to just ahead of them. Nothing that now runs before it
touches optimiser state.
